feat(webapp): centralized password validation for #8
moved password validation logic into /utils/passwords.ts to centralize password validation logic across both the server and the client side
